Refactor email handling: consolidate SMTPCommand and SMTPResponse enums into email.rs, remove parsing module
