+++ src/handler/mod.rs
@@ -1,3 +1,2 @@
 pub mod amqp;
 pub mod email;
-pub mod parsing;
